feat: add owner flow (#2086)
