--- src/components/settings/owner/AddOwnerDialog/DialogSteps/types.d.ts
@@ -1,10 +1,0 @@
-export interface ChangeOwnerData {
-  removedOwner?: OwnerData
-  newOwner: OwnerData
-  threshold?: number
-}
-
-export interface OwnerData {
-  address: string
-  name?: string
-}
