+++ src/components/settings/owner/ReplaceOwnerDialog/index.tsx
@@ -1,11 +1,11 @@
 import { IconButton, Tooltip, SvgIcon } from '@mui/material'
 import { useState } from 'react'
-import { ChooseOwnerStep } from '../AddOwnerDialog/DialogSteps/ChooseOwnerStep'
+// import { ChooseOwnerStep } from '../AddOwnerDialog/DialogSteps/ChooseOwnerStep'
 
 import TxModal from '@/components/tx/TxModal'
 import useSafeInfo from '@/hooks/useSafeInfo'
-import { ReviewOwnerTxStep } from '@/components/settings/owner/AddOwnerDialog/DialogSteps/ReviewOwnerTxStep'
-import type { ChangeOwnerData } from '@/components/settings/owner/AddOwnerDialog/DialogSteps/types'
+// import { ReviewOwnerTxStep } from '@/components/settings/owner/AddOwnerDialog/DialogSteps/ReviewOwnerTxStep'
+// import type { ChangeOwnerData } from '@/components/settings/owner/AddOwnerDialog/DialogSteps/types'
 import type { TxStepperProps } from '@/components/tx/TxStepper/useTxStepper'
 import Track from '@/components/common/Track'
 import { SETTINGS_EVENTS } from '@/services/analytics/events/settings'
@@ -13,14 +13,15 @@ import ReplaceOwnerIcon from '@/public/images/settings/setup/replace-owner.svg'
 import CheckWallet from '@/components/common/CheckWallet'
 
 const ReplaceOwnerSteps: TxStepperProps['steps'] = [
-  {
-    label: 'Replace owner',
-    render: (data, onSubmit) => <ChooseOwnerStep data={data as ChangeOwnerData} onSubmit={onSubmit} />,
-  },
-  {
-    label: 'Review transaction',
-    render: (data, onSubmit) => <ReviewOwnerTxStep data={data as ChangeOwnerData} onSubmit={onSubmit} />,
-  },
+  // TODO: Port in accordance with the new add owner flow
+  // {
+  //   label: 'Replace owner',
+  //   render: (data, onSubmit) => <ChooseOwnerStep data={data as ChangeOwnerData} onSubmit={onSubmit} />,
+  // },
+  // {
+  //   label: 'Review transaction',
+  //   render: (data, onSubmit) => <ReviewOwnerTxStep data={data as ChangeOwnerData} onSubmit={onSubmit} />,
+  // },
 ]
 
 export const ReplaceOwnerDialog = ({ address }: { address: string }) => {
@@ -30,7 +31,7 @@ export const ReplaceOwnerDialog = ({ address }: { address: string }) => {
 
   const { safe } = useSafeInfo()
 
-  const initialModalData: Partial<ChangeOwnerData> = {
+  const initialModalData /*: Partial<ChangeOwnerData>*/ = {
     removedOwner: { address },
     threshold: safe.threshold,
   }
